+++ src/main/java/it/tredi/msa/Msa.java
@@ -1,13 +1,11 @@
 package it.tredi.msa;
 
-import java.security.AccessControlException;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.core.LoggerContext;
 
 /**
  * Classe di definizio e avvio di tutti i job di gestione delle caselle di posta (istanziata da classe Launcher). Si occupa
@@ -16,7 +14,6 @@ import org.apache.logging.log4j.core.LoggerContext;
 public class Msa {
 	
 	private static final Logger logger = LogManager.getLogger(Msa.class.getName());
-	private MsaShutdownHook shutdownHook;
 	
 	private ExecutorServiceHandler executorServiceHandler;
 	
@@ -32,8 +29,6 @@ public class Msa {
 			}
 			throw e;
 		}
-		
-		registerShutdownHook();
 	
 		if (logger.isInfoEnabled())
 			logger.info("MSA Service Started on port [" + port + "]");
@@ -53,41 +48,6 @@ public class Msa {
         executor.shutdown();
 	}
 	
-	protected void registerShutdownHook() {
-		this.shutdownHook = new MsaShutdownHook();
-		try {
-			Runtime.getRuntime().addShutdownHook(shutdownHook);
-			if (logger.isInfoEnabled())
-				logger.info("MsaShutdownHook registered!");
-		} 
-		catch (AccessControlException e) {
-			logger.error("Could not register shutdown hook " + e.getMessage(), e);
-		}		
-	}	
-
-	/**
-	 * Called on shutdown. This gives use a chance to store the keys and to optimize even if the cache manager's shutdown method was not called
-	 * manually.
-	 */
-	class MsaShutdownHook extends Thread {
-
-		/**
-		 * This will persist the keys on shutdown.
-		 * @see java.lang.Thread#run()
-		 */
-		@Override
-		public void run() {
-			if (logger.isInfoEnabled())
-				logger.info("MsaShutdownHook hook ACTIVATED. Shutting down");
-			try {
-				shutdown();
-			} 
-			catch (Exception e) {
-				logger.error("MsaShutdownHook got exception on MSA closure " + e.getMessage(), e);
-			}
-		}
-	}		
-	
 	protected void shutdown() {
 		if (logger.isInfoEnabled())
 			logger.info("shutdown() called");
@@ -96,10 +56,6 @@ public class Msa {
 
 		if (logger.isInfoEnabled())
 			logger.info("shutdown() successfully completed");
-		
-		//call log4j2 shutdown manullay (see log4l2.xml -> shutdownHook="disable")
-		LoggerContext context = (LoggerContext)LogManager.getContext(false);
-		context.stop();
 	}
 
 }