+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [3.0.10] - SNAPSHOT
 
+### Changed
+- Migliorato lo script bash di avvio del servizio su ambiente Linux (controllo di integrità fra PID del processo e file pid memorizzato)
+
 ### Fixed
 - Parsing in DOM di parti di messaggio contenenti caratteri non validi in XML