+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 - Parsing in DOM di parti di messaggio contenenti caratteri non validi in XML
 - In caso di aggiunta di rif. interni ad un documento già registrato (singolo messaggio inviato a 2 mailbox configurate su MSA) occorre aggiornare anche i dati relativi alla storia
+- Data e Ora di invio del messaggio email separata in 2 attributi su XML per problema di chiavi su eXtraWay
 
 
 ## [3.0.9] - 2019-04-18