+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - In caso di aggiunta di rif. interni ad un documento già registrato (singolo messaggio inviato a 2 mailbox configurate su MSA) occorre aggiornare anche i dati relativi alla storia
 - Data e Ora di invio del messaggio email separata in 2 attributi su XML per problema di chiavi su eXtraWay
 - Migliorata la terminazione dell'applicazione: Chiusura delle caselle precedente alla terminazione del contesto Spring (chiusura connessione con MongoDB)
+- Trim su email lette da file di interoperabilità per problemi nell'individuazione del rif. esterno in aggancio di notifiche (Task #17771)
 
 
 ## [3.0.9] - 2019-04-18